modify ResizeMode for TimeSheetView fix bug with Previous button add change frame after change start hour add change frame after change hours & minutes fix bug with move TimeSheetView

+++ OmnieCommerce/Common/MSMPickerViewManager.swift
@@ -191,7 +191,7 @@ extension MSMPickerViewManager: UIPickerViewDelegate {
         case 0:
             if (scene == "TimeSheetPickersView") {
                 selectedHourIndex = row
-                handlerChangeHourCompletion!(row)
+                handlerChangeHourCompletion!(Int16(row))
             } else {
                 selectedDayIndex = row
             }
@@ -209,7 +209,7 @@ extension MSMPickerViewManager: UIPickerViewDelegate {
             
             if (scene == "TimeSheetPickersView") {
                 selectedMinuteIndex = row
-                handlerChangeMinuteCompletion!(row)
+                handlerChangeMinuteCompletion!(Int16(row))
             }
             
         case 4:
